requirements.txt
PyYAML==6.0.2
fastapi
fastapi==0.115.6
httpx==0.27.2
jinja2
langchain
langchain-google-genai
langchain-google-genai==1.0.10
langchain==0.2.16
langgraph
langgraph==0.2.28
openpyxl
openpyxl==3.1.5
pandas
pandas==2.2.3
pydantic-settings==2.5.2
pydantic==2.9.2
python-dotenv
python-dotenv==1.0.1
python-multipart
python-multipart==0.0.12
scikit-learn==1.5.2
uvicorn
uvicorn[standard]==0.30.6
